artifacts/requirements.txt
streamlit>=1.28.0
anthropic>=0.18.0
openai>=1.0.0
python-dotenv>=1.0.0
plotly>=5.17.0
pandas>=2.0.0
jsonschema>=4.19.0